chore: Rename splitDropdown -> SplitDropdown
